requirements.txt
gensim>=3.8.1
jupyter>=1.0.0
matplotlib>=3.0
nltk>=3.4.5
numpy>=1.18
pandas>=1.0
praw>=6.3.1
prawcore>=1.0.1
pyyaml>=5.3.1
scipy>=1.5

